ORコード　endroid/qr-codeからsimplesoftwareio/simple-qrcodeに変更作成中
