+++ docker/php/php.ini
@@ -3,3 +3,8 @@ date.timezone = "Asia/Tokyo"
 [mbstring]
 mbstring.internal_encoding = "UTF-8"
 mbstring.language = "Japanese"
+
+[gd]
+gd.jpeg_ignore_warning = 0
+memory_limit = 256M
+max_execution_time = 300
