+++ src/app/Http/Controllers/BookingController.php
@@ -9,16 +9,16 @@ use App\Models\Favorite;
 use App\Models\Shop;
 use App\Models\User;
 use Illuminate\Support\Facades\Auth;
-use App\Services\QrCodeService;
+// use App\Services\QrCodeService;
 
 class BookingController extends Controller
 {
-  protected $qrCodeService;
+  // protected $qrCodeService;
 
-  public function __construct(QrCodeService $qrCodeService)
-  {
-    $this->qrCodeService = $qrCodeService;
-  }
+  // public function __construct(QrCodeService $qrCodeService)
+  // {
+  //   $this->qrCodeService = $qrCodeService;
+  // }
   /**
    * マイページ全体を表示する.
    *
@@ -130,30 +130,30 @@ class BookingController extends Controller
   }
 
   // 予約時にQRコードを生成するメソッド
-  public function generateQrCode($bookingId)
-  {
-    $booking = Booking::findOrFail($bookingId); // 予約IDで予約情報を取得
-    $token = base64_encode(random_bytes(16)); // 仮のトークンを生成
-    $booking->update(['qr_code_token' => $token]); // 予約情報にトークンを保存
+  // public function generateQrCode($bookingId)
+  // {
+  //   $booking = Booking::findOrFail($bookingId); // 予約IDで予約情報を取得
+  //   $token = base64_encode(random_bytes(16)); // 仮のトークンを生成
+  //   $booking->update(['qr_code_token' => $token]); // 予約情報にトークンを保存
 
-    $qrCodeUrl = route('reservation.scan', ['token' => $token]); // スキャン時にアクセスするURLを生成
-    $qrCodeDataUri = $this->qrCodeService->generate($qrCodeUrl); // 依存注入されたQrCodeServiceを使用してQRコードを生成
+  //   $qrCodeUrl = route('reservation.scan', ['token' => $token]); // スキャン時にアクセスするURLを生成
+  //   $qrCodeDataUri = $this->qrCodeService->generate($qrCodeUrl); // 依存注入されたQrCodeServiceを使用してQRコードを生成
 
-    return view('auth.qr-code', ['qrCode' => $qrCodeDataUri]); // QRコードを表示するビューに渡す
-  }
+  //   return view('auth.qr-code', ['qrCode' => $qrCodeDataUri]); // QRコードを表示するビューに渡す
+  // }
 
   // 店舗側　QRコードのスキャンと認証メソッド
-  public function authenticateReservation(Request $request)
-  {
-    $token = $request->query('token'); // URLからトークンを取得
-    $booking = Booking::where('qr_code_token', $token)->first(); // トークンで予約を検索
+  // public function authenticateReservation(Request $request)
+  // {
+  //   $token = $request->query('token'); // URLからトークンを取得
+  //   $booking = Booking::where('qr_code_token', $token)->first(); // トークンで予約を検索
 
-    if ($booking) {
-      // 予約が見つかった場合、認証成功の処理を実行
-      return redirect()->route('reservation.success');
-    } else {
-      // 予約が見つからない場合、エラーを返す
-      return redirect()->route('reservation.failure');
-    }
-  }
+  //   if ($booking) {
+  //     // 予約が見つかった場合、認証成功の処理を実行
+  //     return redirect()->route('reservation.success');
+  //   } else {
+  //     // 予約が見つからない場合、エラーを返す
+  //     return redirect()->route('reservation.failure');
+  //   }
+  // }
 }
