+++ src/public/css/reservation_failure.css
