+++ src/resources/views/reservation_failure.blade.php
