+++ src/resources/views/reservation_success.blade.php
